+++ docs/explanation/metadata.md
@@ -1,0 +1,5 @@
+# Metadata
+
+explain metadata system, metadata factories 
+
+referenced in ApiResource.php
