+++ docs/pdg.config.json
@@ -1,0 +1,14 @@
+{
+  // Options for the generate-reference.php script 
+  "reference": {
+    "src": "../src"
+  },
+  "sidebar": {
+    "directories": {
+      "Guides": ["pages/guide"],
+      "Reference": ["pages/reference"]
+    },
+    // For URLs
+    "basePath": "pages"
+  }
+}
